Fix bugs found in code review (rounds 4-6)
- Copy radius/direction in sphere/capsule 3-arg translate overloads
- Use constraint distance_threshold_ as spring rest length
- Guard pre_update/intra_update callbacks against null managers
- Fix sphere-vs-convex_solid origin sign (sub → add)
- Only accept entry planes in trace_convex_solid (denom < 0)
- Fix infinite-mass collision response ignoring COR and using wrong impulse
- Handle degenerate sphere-center normal in trace_sphere
- Prevent snap_to_grid UB with int64_t cast

+++ include/hop/scalar_traits.h
@@ -90,7 +90,7 @@ struct scalar_traits<float> {
 
 	// Snap to grid (float version)
 	static void snap_to_grid(float& v, const epsilon_state<float>& s) {
-		v = static_cast<int>((v + (s.half_epsilon * -(v < 0))) * s.one_over_epsilon) * s.epsilon;
+		v = static_cast<int64_t>((v + (s.half_epsilon * -(v < 0))) * s.one_over_epsilon) * s.epsilon;
 	}
 };
